% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionTemplate.R
\name{functionTemplate}
\alias{functionTemplate}
\title{Insert function template.}
\usage{
functionTemplate()
}
\description{
Call this function as an addin to insert a function template at the cursor position.
}
