% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldproxy.R
\name{ldlink_ldproxy}
\alias{ldlink_ldproxy}
\title{1000 Genomes LDproxy}
\usage{
ldlink_ldproxy(rsid = NULL, pop = NULL)
}
\arguments{
\item{rsid}{the rsID for the variant to be searched}

\item{pop}{the 1000 Genomes population short code either super population (e.g EUR, EAS, ...) or sub population (e.g. CEU, CHB, ...)}
}
\description{
A function that will query https://analysistools.nci.nih.gov/LDlink for a proxy SNP.
}
\references{
Machiela MJ, Chanock SJ. LDlink a web-based application for exploring population-specific haplotype structure and linking correlated alleles of possible functional variants. Bioinformatics. 2015 Jul 2. PMID: 26139635.
}
\author{
Murray Cadzow
}
