% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merrimanR.R
\docType{package}
\name{merrimanR}
\alias{merrimanR}
\alias{merrimanR-package}
\title{merrimanR: Functions used in the Merriman Lab.}
\description{
This package provides functions created to facilitate reproducible research
and provide a resource for lab members to be trained in how to create R
packages
}
\section{functions for calculating power}{

powerOR
powerGraph
}

\section{functions for R package teaching examples}{

}

