% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerGraph.R
\name{powerGraph}
\alias{powerGraph}
\title{Create power graphs for case/control GWAS.}
\usage{
powerGraph(caseN = NULL, controlN = NULL, controlA1 = NULL,
  oddsratio = NULL, siglevel = NULL, title = NULL)
}
\arguments{
\item{caseN}{The number of cases}

\item{controlN}{The number of controls}

\item{controlA1}{The allele frequency in controls}

\item{siglevel}{The P value required to be significant}

\item{title}{The title of the graph}
}
\description{
Create power graphs for case/control GWAS.
}
\author{
Tanya Flynn
}
\seealso{
Other power.functions: \code{\link{powerOR}}
}

