% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerOR.R
\name{powerOR}
\alias{powerOR}
\title{Calculate power for case/control GWAS.}
\usage{
powerOR(caseN = NULL, controlN = NULL, controlA1 = NULL,
  oddsratio = NULL, siglevel = NULL)
}
\arguments{
\item{caseN}{The number of cases}

\item{controlN}{The number of controls}

\item{controlA1}{The allele frequency in controls}

\item{siglevel}{The P value required to be significant}

\item{oddratio}{The odds ratio}
}
\description{
Calculate power for case/control GWAS.
}
\author{
Tanya Flynn
}
\seealso{
Other power.functions: \code{\link{powerGraph}}
}

