% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{remove empty columns from data.frame}
\usage{
remove_empty_cols(data)
}
\arguments{
\item{data}{The data.frame to remove empty columns from}
}
\description{
remove empty columns from data.frame
}
\author{
Tanya Flynn
}
