% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fam.R
\name{update_fam}
\alias{update_fam}
\title{Update plink fam file affection column with specified phenotype/affection.}
\usage{
update_fam(fam = NULL, pheno_df = NULL, fid_col = NULL, iid_col = NULL,
  pheno_col = NULL)
}
\arguments{
\item{fam}{The plink fam file read into R as a data.frame}

\item{pheno_df}{The data.frame containing the sample ids and phenotypes}

\item{fid_col}{The column name in pheno_df containing the FIDs}

\item{iid_col}{The column name in pheno_df containing the IIDs}

\item{pheno_col}{The column name in pheno_df containing the phenotypes to put into the fam file}
}
\description{
Update plink fam file affection column with specified phenotype/affection.
}
\author{
Murray Cadzow
}
